refactor(ui): split topology frame assembly
